Move LocalRepositories to coursier module
